move jira commons to own package
