+++ allure-jira-commons/build.gradle
@@ -1,0 +1,22 @@
+description = 'Allure Jira Commons'
+
+apply from: "${gradleScriptDir}/maven-publish.gradle"
+apply from: "${gradleScriptDir}/bintray.gradle"
+apply plugin: 'maven'
+
+dependencies {
+    compileOnly project(':allure-plugin-api')
+    compileOnly('org.projectlombok:lombok')
+
+    compile('com.squareup.retrofit2:retrofit:2.4.0')
+    compile('com.squareup.retrofit2:converter-jackson:2.4.0')
+}
+
+task sourceJar(type: Jar, dependsOn: classes) {
+    classifier = 'sources'
+    from sourceSets.main.allJava
+}
+
+artifacts {
+    archives sourceJar
+}
