+++ plugins/jira-plugin/build.gradle
@@ -5,13 +5,10 @@ apply from: "${gradleScriptDir}/bintray.gradle"
 
 dependencies {
     compileOnly project(':allure-plugin-api')
-    compileOnly('org.projectlombok:lombok')
-
-    compile('com.squareup.retrofit2:retrofit:2.4.0')
-    compile('com.squareup.retrofit2:converter-jackson:2.4.0')
+    compile project(':allure-jira-commons')
 
     testCompile project(':allure-plugin-api')
-
+    
     testCompile('com.github.stefanbirkner:system-rules:1.18.0')
     testCompile('org.mockito:mockito-core')
     testCompile('junit:junit')
